Add LastPricesUpdate in all outcomes
